LayoutTests/imported/w3c/web-platform-tests/css/css-values/calc-size/animation/interpolate-size-height-composition-expected.txt


FAIL Compositing: property <height> underlying [fit-content] from add [100px] to add [200px] at (-0.3) should be [120px] assert_equals: expected "120px " but got "70px "
FAIL Compositing: property <height> underlying [fit-content] from add [100px] to add [200px] at (0) should be [150px] assert_equals: expected "150px " but got "100px "
FAIL Compositing: property <height> underlying [fit-content] from add [100px] to add [200px] at (0.5) should be [200px] assert_equals: expected "200px " but got "150px "
FAIL Compositing: property <height> underlying [fit-content] from add [100px] to add [200px] at (1) should be [250px] assert_equals: expected "250px " but got "200px "
FAIL Compositing: property <height> underlying [fit-content] from add [100px] to add [200px] at (1.5) should be [300px] assert_equals: expected "300px " but got "250px "
FAIL Compositing: property <height> underlying [fit-content] from add [min-content] to add [200px] at (-0.3) should be [5px] assert_equals: expected "5px " but got "50px "
PASS Compositing: property <height> underlying [fit-content] from add [min-content] to add [200px] at (0) should be [50px]
FAIL Compositing: property <height> underlying [fit-content] from add [min-content] to add [200px] at (0.5) should be [125px] assert_equals: expected "125px " but got "200px "
PASS Compositing: property <height> underlying [fit-content] from add [min-content] to add [200px] at (1) should be [200px]
FAIL Compositing: property <height> underlying [fit-content] from add [min-content] to add [200px] at (1.5) should be [275px] assert_equals: expected "275px " but got "200px "
